requirements.txt
boto3
pandas
pyyaml
pytest
